Renaming loginscreen folder, UI for login and registration.
Folder name changed to LoginAndRegistration, made screens for registration and adjusted login screen.
