+++ Dorstip-App-main/.idea/deploymentTargetDropDown.xml
@@ -2,9 +2,6 @@
 <project version="4">
   <component name="deploymentTargetDropDown">
     <value>
-      <entry key="MainActivity">
-        <State />
-      </entry>
       <entry key="app">
         <State />
       </entry>
