+++ Dorstip-App-main/app/src/main/java/com/example/dorstip_app/LoginAndRegistration/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.example.dorstip_app.LoginScreen
+package com.example.dorstip_app.LoginAndRegistration
 
 import android.content.Intent
 import android.os.Bundle
