+++ Dorstip-App-main/app/src/main/java/com/example/dorstip_app/LoginAndRegistration/RegistrationScreen.kt
@@ -1,0 +1,4 @@
+package com.example.dorstip_app.LoginAndRegistration
+
+class RegistrationScreen {
+}
