+++ Dorstip-App-main/app/src/main/res/layout/register_screen.xml
@@ -1,26 +1,24 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
     android:padding="16dp"
     android:gravity="center"
     android:background="@color/background">
 
     <TextView
-        android:id="@+id/tvLoginTitle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/login"
-        android:textSize="36sp"/>
+        android:text="Registration"
+        android:textSize="36dp"
+        />
 
     <EditText
         android:id="@+id/etEmail"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/email"
-        android:inputType="textEmailAddress"
-        />
+        android:inputType="textEmailAddress" />
 
     <EditText
         android:id="@+id/etPassword"
@@ -30,29 +28,16 @@
         android:inputType="textPassword" />
 
     <Button
-        android:layout_marginVertical="10dp"
-        android:id="@+id/btnLogin"
+        android:id="@+id/btnRegister"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:backgroundTint="@color/primary"
-        android:text="@string/login"
-        android:textColor="@color/secondary"/>
+        android:text="@string/register"
+        android:backgroundTint="@color/primary"/>
 
     <TextView
-        android:layout_marginVertical="10dp"
-        android:id="@+id/tvRegister"
+        android:id="@+id/tvLogin"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/no_account_register_here"
-        android:textColor="@color/accent"/>
-
-    <Button
-        android:layout_marginVertical="10dp"
-        android:id="@+id/btnSkip"
-        android:layout_width="133dp"
-        android:layout_height="48dp"
-        android:backgroundTint="@color/primary"
-        android:insetLeft="8dp"
-        android:insetRight="8dp"
-        android:text="@string/skip"/>
+        android:text="@string/already_have_an_account_login_here"
+        android:textColor="@color/accent" />
 </LinearLayout>
