+++ Dorstip-App-main/app/src/main/res/values/strings.xml
@@ -16,4 +16,11 @@
     <string name="tasting_note">Tasting Note</string>
     <string name="description">Description</string>
     <string name="rating">rating</string>
+    <string name="already_have_an_account_login_here">Already have an account? Login here</string>
+    <string name="register">Register</string>
+    <string name="password">Password</string>
+    <string name="email">Email</string>
+    <string name="login">Login</string>
+    <string name="no_account_register_here">No account? Register here</string>
+    <string name="skip">Skip</string>
 </resources>
